+++ ThreeDSecure/src/test/java/com/braintreepayments/api/ThreeDSecureActivityUnitTest.java
@@ -5,6 +5,7 @@ import android.os.Bundle;
 
 import com.cardinalcommerce.cardinalmobilesdk.models.CardinalActionCode;
 import com.cardinalcommerce.cardinalmobilesdk.models.ValidateResponse;
+import com.cardinalcommerce.cardinalmobilesdk.services.CardinalValidateReceiver;
 
 import org.json.JSONException;
 import org.junit.Test;
@@ -13,20 +14,25 @@ import org.mockito.ArgumentCaptor;
 import org.robolectric.RobolectricTestRunner;
 
 import static android.app.Activity.RESULT_OK;
+import static com.braintreepayments.api.ThreeDSecureActivity.RESULT_COULD_NOT_START_CARDINAL;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
+import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Matchers.eq;
 import static org.mockito.Matchers.same;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.never;
 import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
+
+import androidx.fragment.app.FragmentActivity;
 
 @RunWith(RobolectricTestRunner.class)
 public class ThreeDSecureActivityUnitTest {
 
     @Test
-    public void onCreate_invokesCardinalWithLookupData() throws JSONException {
+    public void onCreate_withExtras_invokesCardinalWithLookupData() throws JSONException {
         ThreeDSecureResult threeDSecureResult =
             ThreeDSecureResult.fromJson(Fixtures.THREE_D_SECURE_LOOKUP_RESPONSE);
 
@@ -49,6 +55,26 @@ public class ThreeDSecureActivityUnitTest {
         ThreeDSecureLookup actualLookup = actualResult.getLookup();
         assertEquals("sample-transaction-id", actualLookup.getTransactionId());
         assertEquals("sample-pareq", actualLookup.getPareq());
+    }
+
+    @Test
+    public void onCreate_withoutExtras_finishesWithError() {
+        Intent intent = new Intent();
+        ThreeDSecureActivity sut = spy(new ThreeDSecureActivity());
+        sut.setIntent(intent);
+
+        CardinalClient cardinalClient = mock(CardinalClient.class);
+        sut.onCreateInternal(cardinalClient);
+
+        verify(cardinalClient, never()).continueLookup(any(FragmentActivity.class), any(ThreeDSecureResult.class), any(CardinalValidateReceiver.class));
+        verify(sut).finish();
+
+        ArgumentCaptor<Intent> captor = ArgumentCaptor.forClass(Intent.class);
+        verify(sut).setResult(eq(RESULT_COULD_NOT_START_CARDINAL), captor.capture());
+
+        Intent intentForResult = captor.getValue();
+        assertEquals("Unable to launch 3DS authentication.",
+            intentForResult.getStringExtra(ThreeDSecureActivity.EXTRA_ERROR_MESSAGE));
     }
 
     @Test